fixed mysql floating point

+++ scripts/schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE Probes (
         Id INT NOT NULL AUTO_INCREMENT,
         MeasurementId INT NOT NULL,
         Response VARCHAR(255) NOT NULL,
-        Duration DECIMAL NOT NULL,
+        Duration FLOAT NOT NULL,
         CreatedAt INT NOT NULL,
         PRIMARY KEY(Id),
         CONSTRAINT `fk_measurement_id` FOREIGN KEY (MeasurementId) REFERENCES Measurements (Id) ON DELETE CASCADE ON UPDATE RESTRICT
